[refactor] sync eventlistener 모듈 변경 (order-query-api -> server)
