+++ server/src/main/java/org/flab/deliveryplatform/server/event/outbox/MessageRelayScheduler.java
@@ -14,10 +14,10 @@ import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
 import org.flab.deliveryplatform.common.event.Event;
 import org.flab.deliveryplatform.common.event.EventPublisher;
 import org.flab.deliveryplatform.delivery.interfaces.eventhandler.OrderPayedApplicationEvent;
-import org.flab.deliveryplatform.order.query.interfaces.eventlistener.event.DeliveryCompletedApplicationEvent;
-import org.flab.deliveryplatform.order.query.interfaces.eventlistener.event.MyOrderPayedApplicationEvent;
-import org.flab.deliveryplatform.order.query.interfaces.eventlistener.event.OrderCreatedApplicationEvent;
-import org.flab.deliveryplatform.order.query.interfaces.eventlistener.event.OrderDeliveredApplicationEvent;
+import org.flab.deliveryplatform.server.sync.myorder.eventlistener.event.DeliveryCompletedApplicationEvent;
+import org.flab.deliveryplatform.server.sync.myorder.eventlistener.event.MyOrderPayedApplicationEvent;
+import org.flab.deliveryplatform.server.sync.myorder.eventlistener.event.OrderCreatedApplicationEvent;
+import org.flab.deliveryplatform.server.sync.myorder.eventlistener.event.OrderDeliveredApplicationEvent;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.support.TransactionCallback;
